Make template options clearer #34
